Renamed regex constants to match other modules.
