+++ src/util/RegexConstants.js
